+++ src/ReactSelect/components.tsx
@@ -1,7 +1,9 @@
-import React from "react";
+import React, { createElement } from "react";
 import type {
   MultiValueGenericProps,
   MultiValueRemoveProps,
+  GroupHeadingProps,
+  GroupProps,
 } from "react-select";
 
 import { appendClass } from "src/utils";
@@ -32,3 +34,34 @@ export const MultiValueRemove = ({
 }: MultiValueRemoveProps): JSX.Element => {
   return <span className="icon-close" {...props} />;
 };
+
+export const Group = <Option = unknown, IsMulti extends boolean = boolean>({
+  className,
+  children,
+  Heading,
+  headingProps,
+  innerProps,
+  ...props
+}: GroupProps<Option, IsMulti>) => {
+  return (
+    <div className={`dropdown__group${appendClass(className)}`} {...innerProps}>
+      <Heading {...props} id={headingProps.id}>
+        {headingProps.data.label}
+      </Heading>
+      {children}
+    </div>
+  );
+};
+
+export const GroupHeading = <
+  Option = unknown,
+  IsMulti extends boolean = boolean
+>({
+  className,
+  children,
+  ...props
+}: GroupHeadingProps<Option, IsMulti>) => (
+  <div className={`dropdown__group-header${appendClass(className)}`}>
+    {children}
+  </div>
+);