+++ src/ReactSelect/types.d.ts
@@ -1,8 +1,0 @@
-import type { ReactNode } from "react";
-import type { LabelColor } from "../../Label";
-
-export type CUISelectProps = {
-  label?: ReactNode;
-  multiValueColor?: LabelColor;
-  error?: ReactNode | boolean;
-};